Factor out a module with all the exceptions for clarity and to prevent them residing in weird places to prevent circular imports
