+++ target_postgres/target_tools.py
@@ -9,8 +9,10 @@ import singer
 from singer import utils, metadata, metrics
 
 from target_postgres import json_schema
+from target_postgres.exceptions import TargetError
 from target_postgres.singer_stream import BufferedSingerStream
-from target_postgres.stream_tracker import StreamTracker, TargetError
+from target_postgres.stream_tracker import StreamTracker
+
 
 LOGGER = singer.get_logger()
